UI Improvemnt of City , City View and Country n dwellers long

+++ CityCountryInfoManagement/CityCountryInfoManagement/Gateway/CityGateway.cs
@@ -84,7 +84,7 @@ namespace CityCountryInfoManagement.Gateway
                 city.Name = reader["CityName"].ToString();
                 
                 city.About = reader["CityAbout"].ToString();
-                city.Dwellers = Convert.ToInt32(reader["Dwellers"]);
+                city.Dwellers = Convert.ToInt64(reader["Dwellers"]);
                 city.Location = reader["Location"].ToString();
                 city.Weather = reader["Weather"].ToString();
                 city.CountryId = reader["CountryId"].ToString();
