+++ CityCountryInfoManagement/CityCountryInfoManagement/Models/City.cs
@@ -10,7 +10,7 @@ namespace CityCountryInfoManagement.Models
         public string Name { get; set; }
         public string Location { get; set; }
         public string About { get; set; }
-        public int Dwellers { get; set; }
+        public long Dwellers { get; set; }
         public string Weather { get; set; }
         public string CountryId { get; set; }
         public int Id { get; set; }
